refactor(io): split dataset payload contract from metadata serde
