+++ graphistry/io/types.py
@@ -150,6 +150,27 @@ SIMPLE_ENCODING_CLIENT_KEYS: Tuple[str, ...] = (
 )
 
 
+SIMPLE_ENCODING_SERVER_TO_CLIENT_MAP: Dict[str, str] = {
+    'node_color': 'point_color',
+    'node_size': 'point_size',
+    'node_title': 'point_title',
+    'node_label': 'point_label',
+    'node_icon': 'point_icon',
+    'node_opacity': 'point_opacity',
+    'node_x': 'point_x',
+    'node_y': 'point_y',
+    'edge_color': 'edge_color',
+    'edge_size': 'edge_size',
+    'edge_title': 'edge_title',
+    'edge_label': 'edge_label',
+    'edge_icon': 'edge_icon',
+    'edge_opacity': 'edge_opacity',
+    'edge_source_color': 'edge_source_color',
+    'edge_destination_color': 'edge_destination_color',
+    'edge_weight': 'edge_weight',
+}
+
+
 class NodeEdgeEncodingsDict(TypedDict, total=False):
     """Intermediate encoding structure with bindings and complex encodings.
